Moved graph to internal package
