+++ internal/graph/node.go
